delete unused imports and codes
