+++ judo-psm-generator-jaxrs-impl/src/main/resources/rest/impl/rest.service.impl.java.hbs
@@ -78,9 +78,8 @@ public class {{ implClassName self }} implements {{ namedElementApiRestFqName se
          @io.swagger.v3.oas.annotations.parameters.RequestBody(
                  content = @io.swagger.v3.oas.annotations.media.Content(
                          schema = @io.swagger.v3.oas.annotations.media.Schema(
-                                 //implementation = hu.blackbelt.judo.psm.generator.jaxrs.test.api.actiongrouptest.view.GalaxyDTO.class
                                  {{# if operation.isCreateOperation}}
-                                 implementation = {{ getCreateOperationClassFQName operation }}.class
+                                 implementation = {{ createOperationClassFQName operation }}.class
                                  {{ else }}
                                  implementation = {{ getOperationFQName operation }}.class
                                  {{/ if }}
@@ -93,9 +92,8 @@ public class {{ implClassName self }} implements {{ namedElementApiRestFqName se
         @io.swagger.v3.oas.annotations.parameters.RequestBody(
                 content = @io.swagger.v3.oas.annotations.media.Content(
                         schema = @io.swagger.v3.oas.annotations.media.Schema(
-                                //implementation = hu.blackbelt.judo.psm.generator.jaxrs.test.api.actiongrouptest.view.GalaxyDTO.class
                                  {{# if operation.isCreateOperation}}
-                                 implementation = {{ getCreateOperationClassFQName operation }}.class
+                                 implementation = {{ createOperationClassFQName operation }}.class
                                  {{ else }}
                                  implementation = {{ getOperationFQName operation }}.class
                                  {{/ if }}
